+++ testing/TestingUtils.cpp
@@ -8,32 +8,32 @@
 
 using namespace std;
 
-vector<uint64_t> randomPermutation(size_t size, uint64_t seed) {
+vector<int64_t> randomPermutation(size_t size, int64_t seed) {
     mt19937 gen(seed);
-    vector<uint64_t> retv(size);
+    vector<int64_t> retv(size);
     iota(retv.begin(), retv.end(), 0);
     shuffle(retv.begin(), retv.end(), gen);
     return retv;
 }
 
-vector<uint64_t> randomVec(size_t size, uint64_t low, uint64_t high,
-                           uint64_t seed) {
+vector<int64_t> randomVec(size_t size, int64_t low, int64_t high,
+                          int64_t seed) {
     mt19937 gen(seed);
     return randomVec(size, low, high, gen);
 }
 
-vector<uint64_t> randomVec(size_t size, uint64_t low, uint64_t high,
-                           std::mt19937& gen) {
-    vector<uint64_t> retv(size);
+vector<int64_t> randomVec(size_t size, int64_t low, int64_t high,
+                          std::mt19937& gen) {
+    vector<int64_t> retv(size);
     uniform_int_distribution el(low, high);
-    for (uint64_t i = 0; i < size; i++) {
+    for (int64_t i = 0; i < size; i++) {
         retv[i] = el(gen);
     }
     return retv;
 }
 
 std::vector<double> randomData(size_t size, double low, double high,
-                               uint64_t seed) {
+                               int64_t seed) {
     mt19937 gen(seed);
     return randomData(size, low, high, gen);
 }
@@ -42,32 +42,32 @@ std::vector<double> randomData(size_t size, double low, double high,
                                std::mt19937& gen) {
     vector<double> retv(size);
     uniform_real_distribution<> dis(low, high);
-    for (uint64_t i = 0; i < size; i++) {
+    for (int64_t i = 0; i < size; i++) {
         retv[i] = dis(gen);
     }
     return retv;
 }
 
-vector<uint64_t> randomPartition(size_t weight, uint64_t low, uint64_t high,
-                                 uint64_t seed) {
+vector<int64_t> randomPartition(int64_t weight, int64_t low, int64_t high,
+                                int64_t seed) {
     mt19937 gen(seed);
-    vector<uint64_t> retv;
-    uniform_int_distribution<uint64_t> el(low, high);
+    vector<int64_t> retv;
+    uniform_int_distribution<int64_t> el(low, high);
     while (weight > 0) {
-        uint64_t val = std::min(weight, el(gen));
+        int64_t val = std::min(weight, el(gen));
         retv.push_back(val);
         weight -= val;
     }
     return retv;
 }
 
-string printCols(const vector<set<uint64_t>>& columns) {
+string printCols(const vector<set<int64_t>>& columns) {
     stringstream ss;
     ss << "{\n";
-    for (const set<uint64_t>& col : columns) {
+    for (const set<int64_t>& col : columns) {
         ss << "  { ";
         bool first = true;
-        for (uint64_t c : col) {
+        for (int64_t c : col) {
             ss << (first ? "" : ", ") << c;
             first = false;
         }
@@ -78,13 +78,13 @@ string printCols(const vector<set<uint64_t>>& columns) {
 }
 
 string printPattern(const SparseStructure& mat, bool sym) {
-    uint64_t ord = mat.order();
+    int64_t ord = mat.order();
     vector<bool> isSet(ord * ord, false);
-    for (uint64_t i = 0; i < ord; i++) {
-        uint64_t start = mat.ptrs[i];
-        uint64_t end = mat.ptrs[i + 1];
-        for (uint64_t k = start; k < end; k++) {
-            uint64_t j = mat.inds[k];
+    for (int64_t i = 0; i < ord; i++) {
+        int64_t start = mat.ptrs[i];
+        int64_t end = mat.ptrs[i + 1];
+        for (int64_t k = start; k < end; k++) {
+            int64_t j = mat.inds[k];
             isSet[ord * i + j] = true;  // assume CSR
             if (sym) {
                 isSet[ord * j + i] = true;
@@ -92,8 +92,8 @@ string printPattern(const SparseStructure& mat, bool sym) {
         }
     }
     stringstream ss;
-    for (uint64_t i = 0; i < ord; i++) {
-        for (uint64_t j = 0; j < ord; j++) {
+    for (int64_t i = 0; i < ord; i++) {
+        for (int64_t j = 0; j < ord; j++) {
             ss << (j > 0 ? " " : "") << (isSet[ord * i + j] ? "#" : "_");
         }
         ss << "\n";
@@ -102,26 +102,26 @@ string printPattern(const SparseStructure& mat, bool sym) {
 }
 
 // print sparse structure after collapsing columns according to lumpStart
-string printAggreg(vector<uint64_t> ptrs,  // csc
-                   vector<uint64_t> inds, vector<uint64_t> lumpStart) {
+string printAggreg(vector<int64_t> ptrs,  // csc
+                   vector<int64_t> inds, vector<int64_t> lumpStart) {
     BASPACHO_CHECK_EQ(ptrs.size(), lumpStart.size());
     int64_t ord = lumpStart[lumpStart.size() - 1];
     vector<bool> isSet(ord * ord, false);
-    for (uint64_t i = 0; i < ptrs.size() - 1; i++) {
-        uint64_t start = ptrs[i];
-        uint64_t end = ptrs[i + 1];
-        uint64_t aStart = lumpStart[i];
-        uint64_t aEnd = lumpStart[i + 1];
-        for (uint64_t k = start; k < end; k++) {
-            uint64_t j = inds[k];
-            for (uint64_t a = aStart; a < aEnd; a++) {
+    for (int64_t i = 0; i < ptrs.size() - 1; i++) {
+        int64_t start = ptrs[i];
+        int64_t end = ptrs[i + 1];
+        int64_t aStart = lumpStart[i];
+        int64_t aEnd = lumpStart[i + 1];
+        for (int64_t k = start; k < end; k++) {
+            int64_t j = inds[k];
+            for (int64_t a = aStart; a < aEnd; a++) {
                 isSet[ord * j + a] = true;  // assume CSR
             }
         }
     }
     stringstream ss;
-    for (uint64_t i = 0; i < ord; i++) {
-        for (uint64_t j = 0; j < ord; j++) {
+    for (int64_t i = 0; i < ord; i++) {
+        for (int64_t j = 0; j < ord; j++) {
             ss << (j > 0 ? " " : "") << (isSet[ord * i + j] ? "#" : "_");
         }
         ss << "\n";
@@ -129,13 +129,13 @@ string printAggreg(vector<uint64_t> ptrs,  // csc
     return ss.str();
 }
 
-vector<set<uint64_t>> randomCols(uint64_t size, double fill, uint64_t seed) {
+vector<set<int64_t>> randomCols(int64_t size, double fill, int64_t seed) {
     mt19937 gen(seed);
     uniform_real_distribution<> dis(0.0, 1.0);
-    vector<set<uint64_t>> columns(size);
-    for (uint64_t i = 0; i < size; i++) {
+    vector<set<int64_t>> columns(size);
+    for (int64_t i = 0; i < size; i++) {
         columns[i].insert(i);
-        for (uint64_t j = i + 1; j < size; j++) {
+        for (int64_t j = i + 1; j < size; j++) {
             if (dis(gen) < fill) {
                 columns[i].insert(j);
             }
@@ -144,16 +144,16 @@ vector<set<uint64_t>> randomCols(uint64_t size, double fill, uint64_t seed) {
     return columns;
 }
 
-std::vector<std::set<uint64_t>> joinColums(
-    const std::vector<std::set<uint64_t>>& columns,
-    std::vector<uint64_t> lumpStart) {
+std::vector<std::set<int64_t>> joinColums(
+    const std::vector<std::set<int64_t>>& columns,
+    std::vector<int64_t> lumpStart) {
     BASPACHO_CHECK_EQ(lumpStart[lumpStart.size() - 1], columns.size());
-    std::vector<std::set<uint64_t>> retv;
-    for (uint64_t a = 0; a < lumpStart.size() - 1; a++) {
-        uint64_t start = lumpStart[a];
-        uint64_t end = lumpStart[a + 1];
-        std::set<uint64_t> colz;
-        for (uint64_t i = start; i < end; i++) {
+    std::vector<std::set<int64_t>> retv;
+    for (int64_t a = 0; a < lumpStart.size() - 1; a++) {
+        int64_t start = lumpStart[a];
+        int64_t end = lumpStart[a + 1];
+        std::set<int64_t> colz;
+        for (int64_t i = start; i < end; i++) {
             colz.insert(columns[i].begin(), columns[i].end());
         }
         retv.push_back(colz);
@@ -162,14 +162,14 @@ std::vector<std::set<uint64_t>> joinColums(
 }
 
 // helper
-vector<set<uint64_t>> csrStructToColumns(const SparseStructure& mat) {
-    uint64_t ord = mat.order();
-    vector<set<uint64_t>> columns(ord);
-    for (uint64_t i = 0; i < ord; i++) {
-        uint64_t start = mat.ptrs[i];
-        uint64_t end = mat.ptrs[i + 1];
-        for (uint64_t k = start; k < end; k++) {
-            uint64_t j = mat.inds[k];
+vector<set<int64_t>> csrStructToColumns(const SparseStructure& mat) {
+    int64_t ord = mat.order();
+    vector<set<int64_t>> columns(ord);
+    for (int64_t i = 0; i < ord; i++) {
+        int64_t start = mat.ptrs[i];
+        int64_t end = mat.ptrs[i + 1];
+        for (int64_t k = start; k < end; k++) {
+            int64_t j = mat.inds[k];
             columns[j].insert(i);
         }
     }
@@ -177,9 +177,9 @@ vector<set<uint64_t>> csrStructToColumns(const SparseStructure& mat) {
 }
 
 // helper
-SparseStructure columnsToCscStruct(const vector<set<uint64_t>>& columns) {
-    vector<uint64_t> ptrs, inds;
-    for (const set<uint64_t>& col : columns) {
+SparseStructure columnsToCscStruct(const vector<set<int64_t>>& columns) {
+    vector<int64_t> ptrs, inds;
+    for (const set<int64_t>& col : columns) {
         ptrs.push_back(inds.size());
         inds.insert(inds.end(), col.begin(), col.end());
     }
@@ -188,17 +188,17 @@ SparseStructure columnsToCscStruct(const vector<set<uint64_t>>& columns) {
 }
 
 // naive implementation
-void naiveAddEliminationEntries(vector<set<uint64_t>>& columns, uint64_t start,
-                                uint64_t end) {
+void naiveAddEliminationEntries(vector<set<int64_t>>& columns, int64_t start,
+                                int64_t end) {
     BASPACHO_CHECK_LE(end, columns.size());
     for (int i = start; i < end; i++) {
-        set<uint64_t>& cBlocks = columns[i];
+        set<int64_t>& cBlocks = columns[i];
         auto it = cBlocks.begin();
         BASPACHO_CHECK(it != cBlocks.end());
         BASPACHO_CHECK_EQ(i, *it);  // Expecting diagonal block!;
         while (++it != cBlocks.end()) {
             auto it2 = it;
-            set<uint64_t>& cAltBlocks = columns[*it];
+            set<int64_t>& cAltBlocks = columns[*it];
             while (++it2 != cBlocks.end()) {
                 cAltBlocks.insert(*it2);
             }
@@ -206,15 +206,15 @@ void naiveAddEliminationEntries(vector<set<uint64_t>>& columns, uint64_t start,
     }
 }
 
-vector<set<uint64_t>> makeIndependentElimSet(vector<set<uint64_t>>& columns,
-                                             uint64_t start, uint64_t end) {
-    vector<set<uint64_t>> retvCols(columns.size());
+vector<set<int64_t>> makeIndependentElimSet(vector<set<int64_t>>& columns,
+                                            int64_t start, int64_t end) {
+    vector<set<int64_t>> retvCols(columns.size());
     for (size_t i = 0; i < columns.size(); i++) {
         if (i < start || i >= end) {
             retvCols[i] = columns[i];
         } else {
             retvCols[i].insert(i);
-            for (uint64_t c : columns[i]) {
+            for (int64_t c : columns[i]) {
                 if (c >= end) {
                     retvCols[i].insert(c);
                 }
